restructure components test on separate test modules
